Run simgrid tests through multiprocessing
We need this approach for Must, since in this case we need to parse
the output line after line, to kill the process as soon as a deadlock
is detected (must does not break the deadlock it detects, leading to a
timeouts on the testing infrastructure).

Let's first debug it with simgrid.