CI: Avoid conflicts between the artifacts of all test-must-NN jobs
Before, there were a real possibility that the outcome of one of the
jobs overwrites the outcome of the other ones.

The solution is to not split the runner of a given tool, which is OK
since we preseed the logs with the result of the previous run, so only
new files need to be tested.